3/11/2023

+++ src/main/java/com/example/Library/services/BookService.java
@@ -3,7 +3,6 @@ package com.example.Library.services;
 import com.example.Library.models.Author;
 import com.example.Library.models.Book;
 import com.example.Library.repositories.BookRepository;
-import com.example.Library.util.CheckUtils;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,57 +22,77 @@ public class BookService {
     }
 
     @Transactional
-    public Book addBook(Book book) {
-        CheckUtils.isStringValid(book.getTitle());
-        return bookRepository.save(book);
+    public boolean addBook(Book book) {
+        if (!bookRepository.existsByBookIsbn(book.getIsbn())) {
+            bookRepository.save(book);
+        }
+        return false;
     }
 
-    public List<Book> getBooks() {
+    public List<Book> findAllBook() {
         return bookRepository.findAll();
     }
 
-    public Book getBookById(Long bookId) {
+    public Book findBookById(Long bookId) {
         return bookRepository.findById(bookId).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
+    public List<Book> findBooksByBookTitle(String title) {
+        return bookRepository.findByBookTitle(title).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    }
+
+    public List<Book> findBooksByBookIsbn(String isbn) {
+        return bookRepository.findByBookIsbn(isbn).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    }
+
     @Transactional
-    public void deleteBook(Long bookId) {
-        bookRepository.deleteById(bookId);
+    public boolean deleteBookByBookId(Long bookId) {
+        if (bookRepository.existsById(bookId)) {
+            bookRepository.deleteById(bookId);
+            return true;
+        }
+        return false;
     }
 
-    public List<Book> getAllBook() {
-        return bookRepository.findAll();
+    @Transactional
+    public boolean deleteBookByBookIsbn(String isbn) {
+        if (bookRepository.existsByBookIsbn(isbn)) {
+            bookRepository.deleteByBookIsbn(isbn);
+            return true;
+        }
+        return false;
     }
 
-    public List<Author> getAuthorsByBookId(Long bookId) {
+    public List<Author> findAuthorsByBookId(Long bookId) {
         return bookRepository.findAuthorsByBookId(bookId).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
-    public List<Author> getAuthorsByBookTitle(String title) {
-        return bookRepository.findAuthorsByBookTitle(title).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    public List<Author> findAuthorsByBookIsbn(String isbn) {
+        return bookRepository.findAuthorsByBookIsbn(isbn).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
-    public Date getPublicationDateByBookTitle(String title) {
-        return bookRepository.findPublicationDateByBookTitle(title).orElseThrow(() -> new EntityNotFoundException("Book not found"));
-    }
-
-    public Date getPublicationDateByBookId(Long id) {
+    public Date findPublicationDateByBookId(Long id) {
         return bookRepository.findPublicationDateByBookId(id).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
-    public Integer getTotalQuantityByBookTitle(String title) {
-        return bookRepository.findTotalQuantityByBookTitle(title).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    public Date findPublicationDateByBookIsbn(String isbn) {
+        return bookRepository.findPublicationDateByBookIsbn(isbn).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
-    public Integer getTotalQuantityByBookId(Long id) {
+    public Integer findCurrentQuantityByBookId(Long id) {
+        return bookRepository.findCurrentQuantityByBookId(id).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    }
+
+    public Integer findCurrentQuantityByBookIsbn(String isbn) {
+        return bookRepository.findCurrentQuantityByBookIsbn(isbn).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    }
+
+    public Integer findTotalQuantityByBookId(Long id) {
         return bookRepository.findTotalQuantityByBookId(id).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
-    public Integer getCurrentQuantityByBookTitle(String title) {
-        return bookRepository.findCurrentQuantityByBookTitle(title).orElseThrow(() -> new EntityNotFoundException("Book not found"));
+    public Integer findTotalQuantityByBookIsbn(String isbn) {
+        return bookRepository.findTotalQuantityByBookIsbn(isbn).orElseThrow(() -> new EntityNotFoundException("Book not found"));
     }
 
-    public Integer getCurrentQuantityByBookId(Long id) {
-        return bookRepository.findCurrentQuantityByBookId(id).orElseThrow(() -> new EntityNotFoundException("Book not found"));
-    }
 }
